Bug 1425597 - Use the C++ and rust indexing data from taskcluster.
This gets the most recent searchfox index built on taskcluster
and uses the data from that if possible. If this happens
we can skip doing the mozilla-central build locally.

This is a bit if a conceptual change in how searchfox operates
in that it makes the source repo's CI system responsible for
generating analysis data instead of generating it via local builds.
This will enable searchfox to more easily support more platforms,
branches, and repos. The tradeoff is that the analysis file format
is more exposed and is implicitly part of the API between the
source repo's CI system and the searchfox indexer.