Improve logging in the taskcluster codepath.
This just sprinkles some timestamps into the log so when things
get slow it's easier to identify where the slowness is coming from.
It also cleans up some of the unnecessarily verbose output from
the analysis merging step so that it just outputs one line per
merge instead all of the relevant bash commands.