Bug 1503547 - Reorganize platform-merging code to more easily scale up to more platforms.
Fundamentally this change is a functional no-op. The goal is to organize the
script to do all the per-platform download and prep work in a loop, and the
cross-platform merging stuff after the loop. This makes it much easier to add
new platforms.

Note that there are some minor functional changes. For example, I switched some
uses of wget to curl because the server was returning the JSON file with
'Content-Encoding: gzip' and the only way to reliably decompress that on download
is to use `curl --compressed`.

Also I changed some absolute paths to relative to avoid littering the code with
$INDEX_ROOT everywhere. This makes it a bit easier to read, although it's harder
to rerun failed commands because the paths are relative so you need to be in the
right folder to start with.