Bug 1282123 - Have the mozilla-central indexer also update blame for other branches
The mozilla-central indexer is responsible for uploading the
updated gecko-dev and gecko-blame tarballs back to S3. If we
were to have the mozilla-releases indexer do it as well, we could
run into indexers clobbering each others updates. But this
means that the mozilla-central indexer should also generate
the blame commits needed by the other repos.