Bug 1282123 - Minor cleanup/refactoring of curl/wget usage
Use curl everywhere instead of randomly using wget. Also extract a
couple of variables to avoid repetition. And move the bugzilla-components
download up to the top, so that (a) we fail early on try pushes where
this job is forgotten, which I do all the time and (b) all the
downloads happen before any the processing, which will be useful in the
next few patches.