Bug 1282123 - Parallelize taskcluster downloads.
We spend a lot of time downloading things serially, but it's
trivial to use GNU parallel to parallelize things. So let's do
that. This involves splitting the per-platform loop into two;
the first one generates the download commands, then we run the
downloads, and then we run the second per-platform loop to
process the downloads.