Bug 1595492 - Update mapper downloading.
The mapper files are not .tar.bz2 files with paths that look like:
- build/conversion/nss-dev/.hg/git-mapfile
- ./build/conversion/beagle/.hg/git-mapfile (This is the gecko-dev mapping,
  I'm really hoping beagle is a codename and not the wrong source tree!)

Since we don't care about the paths and there's a single file, we can
use the "-O" option to tar to make it extract the contents to stdout.  So
the tar invocation becomes `-xOj` where x is extract and j is bz2
decompression.

I've tested this locally.