Bug 1619386 - Change the semantics of the second argument to checkout-gecko-repos.sh
This change should be more or less a functional no-op, but sets us up for
the next patch. I say "more or less" because there is one actual effect
from this, which is that the error case triggers an abort earlier (partway
through checkout-gecko-repos.sh, rather than in the <repo>/setup script).
But it doesn't change the conditions under which we error out.