Bug 1621324 - Don't refetch artifacts from taskcluster unless necessary
Before resolving the target gecko revision, this checks to see if
there is an existing set of taskcluster artifacts from a previous run.
Three cases are modified here:
- if a new set of taskcluster artifacts is available, then the new
  INDEXED_HG_REV will be different from PREEXISTING_HG_REV and all
  the old artifacts will be deleted, and new ones downloaded.
- if INDEXED_HG_REV is the same as PREEXISTING_HG_REV, but there are
  platform-specific artifacts available that weren't available before
  (e.g. due to a slow platform-specific build) those will get
  downloaded and the other platform artifacts will get reused.
- if INDEXED_HG_REV is the same as PREEXISTING_HG_REV, but there are
  no new artifacts then all the ones downloaded last time will
  get reused.

In all cases, the intention is that if the previous run aborted/failed
during the setup process, the new run will try that part again. This
is why, for example, checkout-gecko-repos.sh doesn't have an early
exit conditioned on PREEXISTING_HG_REV == INDEXED_HG_REV. If we did
such an early exit, we might avoid doing work that is needed, if the
previous run failed somewhere inside checkout-gecko-repos.sh.

The only scenario that doesn't get recovered that I can think of is
corrupted/partial file downloads, in which case the user has to
manually delete the file (presumably after investigating/verifying
the corruption).