Bug 1641679 - Update gecko-dev branches from the correct remote.
We were updating the branches in gecko-dev from the "origin" remote which we're
no longer pulling updates to. I should have just deleted that remote when I was
migrating to cinnabar, as that would have caused this line to throw an error
which I would have caught. But anyhow, this should fix it so we update from
cinnabar.