Fix process-tc-artifacts.sh for rust 1.47 stdlib.
In rust 1.47 the library is indeed in /library, but there's also
stuff under /src (just `llvm-project/libunwind`).

Since it's not empty we were copying it instead of the rust stdlib.

We could copy both instead, though that might need adding wildcards and
complicating the script. This fixes the indexing error and we can do
that in a follow-up, though it's not clear that having libunwind gives
us much benefit.