Bug 1690181 - Normalize the config4.json repos git/blame folders
This makes it so that all the repos listed in config4.json extract
their git tarballs into a `git` folder and their blame tarballs into
a `blame` folder. This is good for consistency and scriptability.
Also the tarball names are now <repo-name>.tar for the git tarball
and <repo-name-blame>.tar for the blame tarball.