Bug 1690181 - Rename gecko-dev.tar to gecko.tar
All the other repos have their git repo in `<foo>.tar` and
blame repo in `<foo>-blame.tar`. gecko-dev is unique in using
`gecko-dev.tar` and `gecko-blame.tar`. This patch brings gecko
in line with the other repos, by changing the git tarball to
be `gecko.tar`. Note that the folder inside is still `gecko-dev`;
eventually it will be renamed to `git` to match other
repositories but that'll happen when I'm more in the mood to
manually download and re-upload a 10G tarball.