Bug 1712622 - Index the m-c pine project branch. r=kats
I know for "ash" we just called it "ash", but it seems like having the
"mozilla-" prefix better conveys that's it's an m-c branch and that it
uses the same underlying gecko repo/etc. while avoiding the problems
of using "mozilla-central" as a prefix like in "mozilla-central-pine".

I put this in config4 because config4 currently has the fastest indexing
time per my analysis at
https://bugzilla.mozilla.org/show_bug.cgi?id=1712622#c1 and our disk
usage is also relatively low, at 41G / 240G

I'm intentionally leaving this off the help.html listing for now per
discussion with the branch owners.

Because of the implications of the order of the branches in the reblame
list and pine being closed to "central", I've moved pine to always be
the first branch listed after central (or first if there's no central).