Clarify how .cron.yml works
Per discussions in "Firefox CI" on chat.mozilla.org, there is nothing magical about the
mozilla-central `.cron.yml` file.  In the future I'll try and clarify this in the `.cron.yml` file
itself (and potentially prune moot directives like for ESR), but for now I'm updating the
documentation to make it clear here.