Have mozilla-central/setup dynamically create branches
In adding the "pine" branch without this fix, we start from nothing, which
means processing 758,307 commits, which is too many.  This is an attempt to
build on the existing behavior of `reblame` and reduce the number of manual
steps required when creating a new mozilla-central branch.  Previously, this
would have required manual work.