Add a config that's just mozilla-central for expedited testing.
In particular, in cases where I want to update the test expectations
for m-c, it's nice to be able to do this without having to create a new
branch, etc.

I'd previously entertained having a filter mechanism for this, and I
think I may even have patches for that, but this seems more
straightforward in terms of understanding what's happening with tags in
the AWS instance list, etc.