Bug 1759064 - Update checks for multiple source recs
This is from following the steps documented at
https://github.com/mozsearch/mozsearch/blob/master/docs/testing-checks.md#updating-checks-on-a-stopped-aws-indexer-due-to-local-disk-check-failure
after triggering a just-mc run.

I have not added extra checks for the attributes case because my chicken
nuggets are done and I would like to eat them and this check is a pretty
good smoke test check given that it's consistent with our checks in the
tests repo too and all failures would be correlated!