Bug 1783459 - Try switching wubkat to in-tree apt deps, revert to halt
Ubuntu 22.04 has been explicitly supported for ~2 months now it seems per
https://searchfox.org/wubkat/commit/aa960b1a0c7b980efdbc66718239a449b5966a0c
so we should be able to go back to using that.  This will fix current tree
breakage because "unifdef" isn't installed and should ideally generally
improve wubkat stability.

I'm reverting the behavior from "continue" to "halt" but we probably want
an annotation on the root of the config to enable "terminate" rather than
just what our de facto current approach is of "stop" (we shutdown, which
results in the instance being "stopped").