Bug 1786216 - Failover to latest WPT metadata if not present.
The WPT metadata job seems to have a "when" clause that overrides our
request for it to run, so fail over to the "latest" revision if it's
not there.  There is some limited potential for clock skew phenomenon
but it's not huge.

When doing this I noticed the bugzilla components failover was failing
to specify the tree-name for its similar failover so I've addressed
that here too.