Bug 1904890 - Don't walk graphviz submodules.
This also adds just-graphviz.json which matches the
`make build-graphviz-repo` target I added to the Makefile a while ago
that requires it but that I forgot to land.  The rationale for these
files over the jq filtering approach has been that the jq approach
really only works for local indexing runs right now since the indexer
does not know how to perform an equivalent step, but I've often
thought about adding an equivalent step.  That said, the individual
files also allow us to customize the web-server instance type we use
which should usually not be the same tier as if the whole config was
run.