Bug 1858453 - Index mozsearch and mozsearch-mozilla
This adds mozsearch and mozsearch-mozilla to config1 as well as adding
a "just-mozsearch.json" config for local testing.

The "mozsearch" index indexes the clang-plugin, all the python code in
the tree, and does the same rust-analyzer invocation we do for the
"searchfox" config.

The "mozsearch" index does not do the equivalent of
`make build-tests-repo` in this commit; I add that as a separate tree
in the next commit.  I do not do anything to avoid the fulltext search
or the python indexing from indexing into the tests/ tree.