Bug 1923545 - quiet wubkat CMakeScratch warnings, disable docgen again
The Ubuntu upgrade gave us an upgraded cmake version which changed the
path used for `try_compile` per
https://cmake.org/cmake/help/latest/command/try_compile.html#id5 so we
update the ignore path prefix.

I'm also trying to re-disable the documentation generation to improve
build stability and also bisect whether or not that or developer mode
seemed to be the source of our skia problems.