Bug 1900534 - Switch l10n repo to https://github.com/mozilla-l10n/firefox-l10n
I've locally run the steps from
https://github.com/mozsearch/mozsearch/blob/master/docs/newrepo.md for
https://github.com/mozilla-l10n/firefox-l10n and created firefox-l10n git
and blame lz4 tarballs.  I locally have tested that this works by creating a
just-l10n.json config and "build-l10n-repo" target and run with that to
verify I didn't mess up anything obvious, but I won't land those changes
because this is not a tree that justifies having that; we should just do
the dynamic thing so we can specify a config file and a tree in that
config file if we want.