Bug 1890435 - Copy "shared" gecko scripts for forking.
The files are being forked to allow the existing "mozilla-" trees
(and in particular cedar and elm) to continue to work as-is without
risk of breakage from logic changes for the "firefox-" trees.

In order for git to detect the copies by default, it's necessary to touch the
source files.