Bug 1890435 - Switch to github firefox-main for mingw scraping
Yesterday's config4 run failed with a 404 on fetching this file from
hg.mozilla.org.  It seems like this might have been a case where the
apaprent statistical choice not to redirect to hg-edge by the waf
resulted in an overloaded or broken underlying hg server resulted in
a 404, as the file is definitely there when I check and the waf
redirects me to hg-edge.