Bug 1890435 - Remove cypress from, add nss to config4
Cypress is no longer in use as a project repo per its owner, so we can
remove it from config4.  However, note that we are currently leaving it
as one of the branches that the mozilla-central tree updates for now
so we can put it back without to omuch work if it changes.  (This was
confirmed from discussion with :mjf.)

I've moved nss from config1 to config4 because:
- nss was originally useful on config1 as a smoke test because we built
  it for semantic indexing and it it helped stress a number of edge
  cases.  But unfortunately there were build problems on the modern
  version of clang we used and so we had to turn off semantic indexing,
  so it's not really useful to have on config1 as a canary of potential
  problems on other trees.
- nss has broken config1 a number of times; config4 better represents
  the risk/reward profile.